app/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
Werkzeug==3.0.6
zipp==3.20.2